requirements.txt
# PDF to SEIR Model Generator Requirements

# Core AI/ML
google-generativeai>=0.3.0
python-dotenv>=0.19.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Standard libraries (usually included with Python)
# xml.etree.ElementTree - built-in
# json - built-in  
# re - built-in
# os - built-in
# datetime - built-in